Fix websocket message handler
Fix websocket message handler when the message is empty.
Fix the killing of the Containerzier.

+++ Containerizer/Containerizer.Tests/Specs/Helpers.cs
@@ -6,6 +6,7 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Net;
 using System.Net.Http;
 using System.Reflection;
@@ -135,20 +136,29 @@ namespace Containerizer.Tests.Specs
 
         public static ContainerizerProcess CreateContainerizerProcess()
         {
-            foreach (Process p in System.Diagnostics.Process.GetProcessesByName("Containerizer.exe"))
+            while (true)
             {
-                try
+                var containzerizerProcesses = Process.GetProcessesByName("Containerizer");
+                if (containzerizerProcesses.Length == 0)
                 {
-                    p.Kill();
-                    p.WaitForExit(); // possibly with a timeout
+                    break;
                 }
-                catch (Win32Exception winException)
+
+                foreach (Process p in containzerizerProcesses)
                 {
-                    // process was terminating or can't be terminated - deal with it
-                }
-                catch (InvalidOperationException invalidException)
-                {
-                    // process has already exited - might be able to let this one go
+                    try
+                    {
+                        p.Kill();
+                        p.WaitForExit(); // possibly with a timeout
+                    }
+                    catch (Win32Exception winException)
+                    {
+                        // process was terminating or can't be terminated - deal with it
+                    }
+                    catch (InvalidOperationException invalidException)
+                    {
+                        // process has already exited - might be able to let this one go
+                    }
                 }
             }
             var port = 48080;
